--- apps/backend-workers/.env.example
@@ -1,7 +1,0 @@
-OPENAI_API_KEY=
-
-USE_OPENAI_MODEL_NAME=
-
-# Better-auth
-BETTER_AUTH_SECRET=
-BETTER_AUTH_URL=http://localhost:8787 # Base URL of your app
