--- apps/backend-workers/.gitignore
@@ -1,42 +1,0 @@
-# prod
-dist/
-
-# dev
-.yarn/
-!.yarn/releases
-.vscode/*
-!.vscode/launch.json
-!.vscode/*.code-snippets
-.idea/workspace.xml
-.idea/usage.statistics.xml
-.idea/shelf
-
-# deps
-node_modules/
-.wrangler
-
-# env
-.env
-.env.production
-.dev.vars
-
-# logs
-logs/
-*.log
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-pnpm-debug.log*
-lerna-debug.log*
-
-# misc
-.DS_Store
-
-/generated/prisma
-
-
-# testに関する設定
-/testImage
-
-# test/__mocks__/cloudflareWorkersMock.ts
-**/test/__mocks__/cloudflareWorkersMock.ts
