--- apps/backend-workers/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "backend-workers",
-  "type": "module",
-  "scripts": {
-    "dev": "wrangler dev --port 8787 --ip 0.0.0.0",
-    "deploy": "wrangler deploy --minify",
-    "cf-typegen": "wrangler types --env-interface CloudflareBindings",
-    "local:migration": "wrangler d1 migrations apply buildcha --local",
-    "remote:migration": "wrangler d1 migrations apply buildcha --remote",
-    "test": "jest"
-  },
-  "dependencies": {
-    "@hono/zod-validator": "^0.7.2",
-    "@langchain/cloudflare": "^0.1.1",
-    "@langchain/core": "^0.3.75",
-    "@langchain/openai": "^0.6.9",
-    "@prisma/adapter-d1": "^6.15.0",
-    "@prisma/client": "^6.15.0",
-    "@scalar/hono-api-reference": "^0.9.18",
-    "better-auth": "^1.3.8",
-    "hono": "^4.9.6",
-    "hono-openapi": "0.4.8",
-    "zod": "^4.0.17",
-    "zod-openapi": "^5.3.1"
-  },
-  "devDependencies": {
-    "@cloudflare/workers-types": "^4.20250912.0",
-    "@types/jest": "^30.0.0",
-    "jest": "^30.1.3",
-    "prisma": "^6.15.0",
-    "ts-jest": "^29.4.1",
-    "wrangler": "^4.4.0"
-  }
-}
