--- apps/backend-workers/src/moc/getAnswerObject.ts
@@ -1,17 +1,0 @@
-//とりあえずMocを作成
-export const getAnswerObjectImageUrls = (_questId: string) => {
-  return {
-    topView:
-      "https://www.silhouette-illust.com/wp-content/uploads/2016/10/15213-600x600.jpg",
-    bottomView:
-      "https://www.silhouette-illust.com/wp-content/uploads/2016/10/15213-600x600.jpg",
-    leftView:
-      "https://www.silhouette-illust.com/wp-content/uploads/2016/10/15213-600x600.jpg",
-    rightView:
-      "https://www.silhouette-illust.com/wp-content/uploads/2016/10/15213-600x600.jpg",
-    frontView:
-      "https://www.silhouette-illust.com/wp-content/uploads/2016/10/15213-600x600.jpg",
-    backView:
-      "https://www.silhouette-illust.com/wp-content/uploads/2016/10/15213-600x600.jpg",
-  };
-};
