+++ apps/backend/jest.config.js
@@ -1,11 +1,8 @@
-const { createDefaultPreset } = require("ts-jest");
-
-const tsJestTransformCfg = createDefaultPreset().transform;
-
-/** @type {import("jest").Config} **/
-module.exports = {
-  testEnvironment: "node",
-  transform: {
-    ...tsJestTransformCfg,
+export default {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  testMatch: ['**/test/**/*.test.ts'],
+  moduleNameMapper: {
+    '^cloudflare:workers$': '<rootDir>/test/__mocks__/cloudflareWorkersMock.ts',
   },
-};
+}
