+++ apps/backend/package.json
@@ -1,44 +1,34 @@
 {
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "dist/src/functions/*.js",
+  "name": "backend-workers",
+  "type": "module",
   "scripts": {
-    "build": "tsc",
-    "watch": "tsc -w",
-    "clean": "rimraf dist",
-    "prestart": "pnpm run clean && pnpm run build",
-    "start": "func start",
-    "dev": "tsx watch src/local.ts",
-    "migrate": "prisma migrate dev --schema prisma/schema.prisma",
+    "dev": "wrangler dev --port 8787 --ip 0.0.0.0",
+    "deploy": "wrangler deploy --minify",
+    "cf-typegen": "wrangler types --env-interface CloudflareBindings",
+    "local:migration": "wrangler d1 migrations apply buildcha --local",
+    "remote:migration": "wrangler d1 migrations apply buildcha --remote",
     "test": "jest"
   },
   "dependencies": {
-    "@azure/functions": "^4.0.0",
-    "@hono/node-server": "^1.17.1",
-    "@hono/swagger-ui": "^0.5.2",
     "@hono/zod-validator": "^0.7.2",
-    "@langchain/core": "^0.3.66",
+    "@langchain/cloudflare": "^0.1.1",
+    "@langchain/core": "^0.3.75",
     "@langchain/openai": "^0.6.9",
-    "@marplex/hono-azurefunc-adapter": "^1.0.1",
-    "dotenv": "^17.2.1",
-    "hono": "^4.8.4",
+    "@prisma/adapter-d1": "^6.15.0",
+    "@prisma/client": "^6.15.0",
+    "@scalar/hono-api-reference": "^0.9.18",
+    "better-auth": "^1.3.8",
+    "hono": "^4.9.6",
     "hono-openapi": "0.4.8",
-    "langchain": "^0.3.30",
-    "prisma": "^6.14.0",
-    "tsx": "^4.20.3",
     "zod": "^4.0.17",
     "zod-openapi": "^5.3.1"
   },
   "devDependencies": {
-    "@prisma/client": "^6.14.0",
+    "@cloudflare/workers-types": "^4.20250912.0",
     "@types/jest": "^30.0.0",
-    "@types/node": "18.x",
-    "@types/swagger-ui-dist": "^3.30.6",
-    "concurrently": "^9.2.0",
-    "jest": "^30.1.2",
-    "rimraf": "^5.0.0",
+    "jest": "^30.1.3",
+    "prisma": "^6.15.0",
     "ts-jest": "^29.4.1",
-    "typescript": "^5.8.3"
+    "wrangler": "^4.4.0"
   }
 }
