+++ apps/backend/src/lib/auth.ts
